1/requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-cloud-vision==3.7.1
google-cloud-speech==2.24.1
google-cloud-texttospeech==2.15.0
google-generativeai==0.3.2
python-dotenv==1.0.1
PyPDF2==3.0.1